move converters under distance namespace;
